[nomaster] can't add new class BatchingExecutor
This is necessary to maintain binary compatibility with 2.10.0.
