+++ bincompat-forward.whitelist.conf
@@ -34,16 +34,16 @@ filter {
         #     matchName="scala.util.Random$"
         #     problemName=MissingTypesProblem
         # },
-        {
-            # private[concurrent]
-            matchName="scala.concurrent.BatchingExecutor$Batch"
-            problemName=MissingClassProblem
-        },
-        {
-            # private[concurrent]
-            matchName="scala.concurrent.BatchingExecutor"
-            problemName=MissingClassProblem
-        },
+        # {
+        #     # private[concurrent]
+        #     matchName="scala.concurrent.BatchingExecutor$Batch"
+        #     problemName=MissingClassProblem
+        # },
+        # {
+        #     # private[concurrent]
+        #     matchName="scala.concurrent.BatchingExecutor"
+        #     problemName=MissingClassProblem
+        # },
         {
             # private[concurrent]
             matchName="scala.concurrent.impl.ExecutionContextImpl$AdaptedForkJoinTask"
